// src/types/resources/RelatedPerson.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { THumanName } from '../partials/HumanName';
import { TContactPoint } from '../partials/ContactPoint';
import { TDate } from '../simpleTypes/Date';
import { TAddress } from '../partials/Address';
import { TAttachment } from '../partials/Attachment';
import { TPeriod } from '../partials/Period';
import { TRelatedPersonCommunication } from '../partials/RelatedPersonCommunication';

export type TRelatedPerson = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  active?: boolean;
  patient: TReference;
  relationship?: TCodeableConcept[];
  name?: THumanName[];
  telecom?: TContactPoint[];
  gender?: string;
  birthDate?: TDate;
  address?: TAddress[];
  photo?: TAttachment[];
  period?: TPeriod;
  communication?: TRelatedPersonCommunication[];
};